split phylesystem.__init__.py into separate files
This was long overdue. It had utility functions (now
in phylesystem.helper), shard operations (now in
phylesystem/phylesystem_shard), and operations that
act on the wrapper around all of the shards (now in
phylesystem/phylesystem_umbrella).

I updated the imports so that they would refer to the
new location. But __ini__.py imports the high-level
Phylesystem class (from the umbrella file). So client
code (like the phylesystem-api repo) won't need to change
what they import.
