+++ peyotl/test/support/pathmap.py
@@ -71,7 +71,7 @@ def get_test_phylesystem_mirror_info():
     return {'push': {'parent_dir': get_test_phylesystem_mirror_parent()}}
 
 def get_test_phylesystem():
-    from peyotl.phylesystem import _Phylesystem
+    from peyotl.phylesystem.phylesystem_umbrella import _Phylesystem
     r = get_test_repos()
     mi = get_test_phylesystem_mirror_info()
     mi['push']['remote_map'] = {'GitHubRemote': 'git@github.com:mtholder'}
