+++ peyotl/test/test_tmp_file_write.py
@@ -3,7 +3,7 @@ from peyotl.phylesystem.git_actions import GitAction
 import unittest
 from peyotl.nexson_syntax import read_as_json
 from peyotl.test.support import pathmap
-from peyotl.phylesystem import get_repos
+from peyotl.phylesystem.helper import get_repos
 try:
     r = get_repos()
     HAS_LOCAL_PHYLESYSTEM_REPOS = True
